survey input sort column data type change to unsigned integer

+++ database/migrations/2016_11_13_085916_create_survey_inputs_table.php
@@ -22,7 +22,7 @@ class CreateSurveyInputsTable extends Migration
             $table->string('value')->index();
             $table->string('className')->nullable();
             $table->string('skip')->nullable();
-            $table->unsignedSmallInteger('sort')->index();
+            $table->unsignedInteger('sort')->index();
             $table->unsignedTinyInteger('section')->index();
             $table->enum('status', ['new', 'published'])->default('new');
             $table->boolean('double_entry')->default(false)->nullable();
